+++ tests/.eslintrc.js
@@ -1,5 +1,12 @@
 module.exports = {
   env: {
     embertest: true
+  },
+  globals: {
+    introJSEnsureClosed: true,
+    introJSCurrentStep: true,
+    introJSExit: true,
+    introJSNext: true,
+    introJSPrevious: true
   }
 };
