+++ tests/dummy/app/controllers/application.js
@@ -1,0 +1,14 @@
+import Ember from 'ember';
+
+export default Ember.Controller.extend({
+  steps: [
+    {
+      intro: 'Step 1!',
+      element: '#step1'
+    },
+    {
+      intro: 'Step 2!',
+      element: '#step2'
+    }
+  ]
+});
