+++ tests/dummy/config/environment.js
@@ -42,9 +42,5 @@ module.exports = function(environment) {
     ENV.APP.rootElement = '#ember-testing';
   }
 
-  if (environment === 'production') {
-
-  }
-
   return ENV;
 };
