+++ tests/test-helper.js
@@ -1,4 +1,6 @@
 import resolver from './helpers/resolver';
 import { setResolver } from 'ember-mocha';
 
+import './helpers/ember-introjs';
+
 setResolver(resolver);
